LayoutTests/fast/regions/get-region-flow-ranges-absolute-pos.html

<!doctype html>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script src="resources/helper.js"></script>
<style>
    .border { border: 1px solid red; }
    #box1 { -webkit-flow-into: flow; position: absolute; width: 50px; height: 50px; background-color: green; top: 25px; left: 25px; }
    #region1 { -webkit-flow-from: flow; width: 100px; height: 100px; position: absolute; top: 100px; left: 50px; }
</style>
</head>
<body>
    <div id="description"></div>
    <div id="wrapper"><div id="box1"></div></div>
    <div id="region1" class="border"></div>
<script>
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    debug("[CSSRegions] Test for Region::getRegionFlowRanges for absolute positioned elements");
    var region1 = document.getElementById("region1");
    range1 = region1.webkitGetRegionFlowRanges();
    shouldEvaluateTo("range1.length", 1);
    compareArrays(getRangeAt(range1, 0), ["wrapper", 0, "wrapper", 1]);

function hideRegionsAndFlows() {
    document.getElementById("region1").style.visibility = "hidden";
}

    hideRegionsAndFlows();
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
